+++ src/components/Header/Header.jsx
@@ -1,0 +1,16 @@
+import React from 'react'
+import { Link } from 'react-router-dom';
+
+const Header = () => {
+  return (
+    <div>
+      <ul>
+        <li><Link to='/'>Home</Link></li>
+        <li><Link to='/signin'>Sign In</Link></li>
+        <li><Link to='/signup'>Sign Up</Link></li>
+      </ul>
+    </div>
+  )
+}
+
+export default Header
