+++ src/components/Home/Home.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Home = () => {
+  return (
+    <div>
+      <h2>Home</h2>
+    </div>
+  )
+}
+
+export default Home
