+++ src/components/SignUp/SignUp.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const SignUp = () => {
+  return (
+    <div>
+      <h2>SignUp</h2>
+    </div>
+  )
+}
+
+export default SignUp
